Pull draw functions out to clear up diff.
