+++ bevy_nannou_render/src/pipeline.rs
@@ -1,42 +1,28 @@
 use std::hash::Hash;
 
-use bevy::core_pipeline::core_3d::{Transparent3d, CORE_3D_DEPTH_FORMAT};
-use bevy::ecs::query::ROQueryItem;
-use bevy::ecs::system::lifetimeless::{Read, SRes};
-use bevy::ecs::system::SystemParamItem;
-use bevy::pbr::{MeshPipeline, MeshPipelineKey, SetMeshViewBindGroup};
+use bevy::core_pipeline::core_3d::CORE_3D_DEPTH_FORMAT;
+use bevy::pbr::{MeshPipeline, MeshPipelineKey};
 use bevy::prelude::*;
-use bevy::render::extract_component::DynamicUniformIndex;
-use bevy::render::render_asset::RenderAssets;
-use bevy::render::render_phase::{
-    DrawFunctions, PhaseItem, RenderCommand, RenderCommandResult, RenderPhase, SetItemPipeline,
-    TrackedRenderPass,
-};
+use bevy::render::render_phase::RenderCommand;
 use bevy::render::render_resource as wgpu;
-use bevy::render::render_resource::{
-    PipelineCache, RenderPipelineDescriptor, SpecializedRenderPipeline, SpecializedRenderPipelines,
-};
+use bevy::render::render_resource::{RenderPipelineDescriptor, SpecializedRenderPipeline};
 use bevy::render::renderer::RenderDevice;
-use bevy::render::texture::BevyDefault;
-use bevy::render::view::{ExtractedView, ViewTarget};
+use bevy::render::view::ViewTarget;
 
 use bevy_nannou_draw::draw::mesh;
 use bevy_nannou_draw::draw::mesh::vertex::Point;
 
-use crate::{
-    DefaultTextureHandle, DrawMesh, DrawMeshHandle, DrawMeshItem, DrawMeshUniform,
-    DrawMeshUniformBindGroup, TextureBindGroupCache, NANNOU_SHADER_HANDLE,
-};
+use crate::{DrawMeshUniform, NANNOU_SHADER_HANDLE};
 
 #[derive(Resource)]
 pub struct NannouPipeline {
-    mesh_pipeline: MeshPipeline,
-    glyph_cache_texture: wgpu::Texture,
-    text_bind_group_layout: wgpu::BindGroupLayout,
-    text_bind_group: wgpu::BindGroup,
-    pub(crate) texture_bind_group_layout: wgpu::BindGroupLayout,
-    texture_bind_group: wgpu::BindGroup,
-    pub(crate) mesh_bind_group_layout: wgpu::BindGroupLayout,
+    pub mesh_pipeline: MeshPipeline,
+    pub glyph_cache_texture: wgpu::Texture,
+    pub text_bind_group_layout: wgpu::BindGroupLayout,
+    pub text_bind_group: wgpu::BindGroup,
+    pub texture_bind_group_layout: wgpu::BindGroupLayout,
+    pub texture_bind_group: wgpu::BindGroup,
+    pub mesh_bind_group_layout: wgpu::BindGroupLayout,
 }
 
 // This key is computed and used to cache the pipeline.
@@ -247,183 +233,13 @@ impl FromWorld for NannouPipeline {
         let mesh_pipeline = render_world.resource::<MeshPipeline>().clone();
 
         NannouPipeline {
+            mesh_pipeline,
+            mesh_bind_group_layout,
             glyph_cache_texture,
             text_bind_group_layout,
             text_bind_group,
             texture_bind_group_layout,
             texture_bind_group,
-            mesh_bind_group_layout,
-            mesh_pipeline,
         }
-    }
-}
-
-pub fn queue_draw_mesh_items(
-    draw_functions: Res<DrawFunctions<Transparent3d>>,
-    pipeline: Res<NannouPipeline>,
-    mut pipelines: ResMut<SpecializedRenderPipelines<NannouPipeline>>,
-    pipeline_cache: Res<PipelineCache>,
-    msaa: Res<Msaa>,
-    items: Query<(Entity, &DrawMeshItem)>,
-    mut views: Query<(&ExtractedView, &mut RenderPhase<Transparent3d>)>,
-) {
-    let draw_function = draw_functions
-        .read()
-        .get_id::<DrawDrawMeshItem3d>()
-        .unwrap();
-    for (view, mut transparent_phase) in &mut views {
-        for (entity, item) in items.iter() {
-            let key = NannouPipelineKey {
-                output_color_format: if view.hdr {
-                    ViewTarget::TEXTURE_FORMAT_HDR
-                } else {
-                    wgpu::TextureFormat::bevy_default()
-                },
-                sample_count: msaa.samples(),
-                topology: item.topology,
-                blend_state: item.blend,
-            };
-
-            let pipeline = pipelines.specialize(&pipeline_cache, &pipeline, key);
-
-            transparent_phase.add(Transparent3d {
-                entity,
-                draw_function,
-                pipeline,
-                distance: 0.,
-                batch_range: 0..1,
-                dynamic_offset: None,
-            });
-        }
-    }
-}
-
-pub type DrawDrawMeshItem3d = (
-    SetItemPipeline,
-    SetMeshViewBindGroup<0>,
-    SetDrawMeshUniformBindGroup<1>,
-    SetDrawMeshTextBindGroup<2>,
-    SetDrawMeshTextureBindGroup<3>,
-    SetDrawMeshScissor,
-    DrawDrawMeshItem,
-);
-
-pub struct SetDrawMeshUniformBindGroup<const I: usize>;
-impl<const I: usize, P: PhaseItem> RenderCommand<P> for SetDrawMeshUniformBindGroup<I> {
-    type Param = SRes<DrawMeshUniformBindGroup>;
-    type ViewWorldQuery = ();
-    type ItemWorldQuery = Read<DynamicUniformIndex<DrawMeshUniform>>;
-
-    #[inline]
-    fn render<'w>(
-        _item: &P,
-        _view: ROQueryItem<'w, Self::ViewWorldQuery>,
-        uniform_index: ROQueryItem<'w, Self::ItemWorldQuery>,
-        bind_group: SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        pass.set_bind_group(
-            I,
-            &bind_group.into_inner().bind_group,
-            &[uniform_index.index()],
-        );
-        RenderCommandResult::Success
-    }
-}
-
-pub struct SetDrawMeshTextureBindGroup<const I: usize>;
-impl<const I: usize, P: PhaseItem> RenderCommand<P> for SetDrawMeshTextureBindGroup<I> {
-    type Param = (SRes<DefaultTextureHandle>, SRes<TextureBindGroupCache>);
-    type ViewWorldQuery = ();
-    type ItemWorldQuery = Read<DrawMeshItem>;
-
-    #[inline]
-    fn render<'w>(
-        _item: &P,
-        _view: ROQueryItem<'w, Self::ViewWorldQuery>,
-        draw_mesh_item: ROQueryItem<'w, Self::ItemWorldQuery>,
-        (default_texture, bind_groups): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let texture = match &draw_mesh_item.texture {
-            None => &default_texture.0,
-            Some(texture) => texture,
-        };
-
-        let Some(bind_group) = bind_groups.into_inner().get(texture) else {
-            return RenderCommandResult::Failure;
-        };
-
-        pass.set_bind_group(I, &bind_group, &[]);
-        RenderCommandResult::Success
-    }
-}
-
-pub struct SetDrawMeshTextBindGroup<const I: usize>;
-impl<const I: usize, P: PhaseItem> RenderCommand<P> for SetDrawMeshTextBindGroup<I> {
-    type Param = SRes<NannouPipeline>;
-    type ViewWorldQuery = ();
-    type ItemWorldQuery = ();
-
-    #[inline]
-    fn render<'w>(
-        _item: &P,
-        _view: ROQueryItem<'w, Self::ViewWorldQuery>,
-        _entity: ROQueryItem<'w, Self::ItemWorldQuery>,
-        pipeline: SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        pass.set_bind_group(I, &pipeline.into_inner().text_bind_group, &[]);
-        RenderCommandResult::Success
-    }
-}
-
-pub struct SetDrawMeshScissor;
-impl<P: PhaseItem> RenderCommand<P> for SetDrawMeshScissor {
-    type Param = ();
-    type ViewWorldQuery = ();
-    type ItemWorldQuery = Read<DrawMeshItem>;
-
-    fn render<'w>(
-        _item: &P,
-        _view: ROQueryItem<'w, Self::ViewWorldQuery>,
-        entity: ROQueryItem<'w, Self::ItemWorldQuery>,
-        _param: SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        if let Some(scissor) = entity.scissor {
-            pass.set_scissor_rect(scissor.left, scissor.bottom, scissor.width, scissor.height);
-        }
-
-        RenderCommandResult::Success
-    }
-}
-
-pub struct DrawDrawMeshItem;
-impl<P: PhaseItem> RenderCommand<P> for DrawDrawMeshItem {
-    type Param = SRes<RenderAssets<DrawMesh>>;
-    type ViewWorldQuery = Read<DrawMeshHandle>;
-    type ItemWorldQuery = Read<DrawMeshItem>;
-
-    #[inline]
-    fn render<'w>(
-        _item: &P,
-        handle: ROQueryItem<'w, Self::ViewWorldQuery>,
-        draw_mesh_item: ROQueryItem<'w, Self::ItemWorldQuery>,
-        draw_meshes: SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let Some(mesh) = draw_meshes.into_inner().get(&handle.0) else {
-            return RenderCommandResult::Failure;
-        };
-
-        // Set the buffers.
-        pass.set_index_buffer(mesh.index_buffer.slice(..), 0, wgpu::IndexFormat::Uint32);
-        pass.set_vertex_buffer(0, mesh.point_buffer.slice(..));
-        pass.set_vertex_buffer(1, mesh.color_buffer.slice(..));
-        pass.set_vertex_buffer(2, mesh.tex_coords_buffer.slice(..));
-
-        pass.draw_indexed(draw_mesh_item.index_range.clone(), 0, 0..1);
-        RenderCommandResult::Success
     }
 }
